Bugfixes and moving to WKT serialization of geometry data
- DataFrameCache now uses its own sequence number attribute to avoid breaking rowid
- Fixed a crasher in UNIONs (ROWIDs were being stored alternatively as LongType and IntegerType)
- SparkPrimitive now serializes geometry data as WKT rather than a Base64-encoded array.

+++ src/main/scala/org/mimirdb/rowids/AnnotateWithRowIds.scala
@@ -152,7 +152,7 @@ class AnnotateWithRowIds(
           Union(
             children.zipWithIndex.map { case (child, idx) => 
               val (rewrite, annotation) = recur(child)
-              annotate(rewrite, newAnnotation.exprId, Literal(idx), annotation)
+              annotate(rewrite, newAnnotation.exprId, Literal(idx.toLong), annotation)
             }
           )
         (ret, newAnnotation)
